trekk ut loggmetrikker i egen fil og legg til metrikker for AxsysRestClient
